Setting Page: Code Optimized & Auth Gate Fixed
